Fix type

+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/dataset/DataSet.java
@@ -635,7 +635,7 @@ public class DataSet extends BaseDimensionalItemObject
     return timelyDays;
   }
 
-  public void setTimelyDays(int timelyDays) {
+  public void setTimelyDays(double timelyDays) {
     this.timelyDays = timelyDays;
   }
